fix: prevent nones
