+++ backend/src/repositories/resolver/person_resolver.py
@@ -90,7 +90,7 @@ class BasicPersonResolver(AbstractResolver[Person]):
                 )
             )
 
-        birth_date = entity.biography.birth_date
+        birth_date = entity.biography.birth_date if entity.biography else None
         if birth_date:
             # parse date_naissance
             ddp = DateDataParser(languages=["fr"])
@@ -110,7 +110,7 @@ class BasicPersonResolver(AbstractResolver[Person]):
                 chat = OllamaDateFormatter(settings=self.settings)
                 birth_date = chat.format(birth_date)
 
-        death_date = entity.biography.death_date
+        death_date = entity.biography.death_date if entity.biography else None
         if death_date:
             # parse date_deces
             ddp = DateDataParser(languages=["fr"])
