+++ backend/tests/repositories/ml/test_ollama_rag.py
@@ -1,10 +1,9 @@
-from pydantic import BaseModel, Field, HttpUrl
+from pydantic import HttpUrl
 
 from src.entities.extraction import ExtractionResult
 from src.entities.person import Biography
 from src.entities.source import SourcedContentBase
 from src.repositories.ml.ollama_generic import GenericInfoExtractor
-from src.repositories.ml.response_formater import create_response_model
 from src.settings import Settings
 
 
@@ -57,45 +56,3 @@ def test_ollama_is_called_correctly(mocker):
     assert isinstance(result.entity, Biography)
     assert result.entity.full_name == "Quentin Jerome Tarantino"
     assert result.score == 0.95
-
-
-def test_create_response_model():
-
-    # given
-    class MyModel(BaseModel):
-        height: int = Field(..., description="Height in centimeters")
-
-    # when
-    response = create_response_model(MyModel)(score=0.9, height=180)
-
-    # then
-    assert response.score == 0.9
-    assert response.height == 180
-
-
-def test_create_response_model_excludes_http_fields():
-
-    # given
-    class MyModel(BaseModel):
-        name: str = Field(..., description="Name of the person")
-        profile_url: HttpUrl = Field(..., description="Profile URL")
-        list_of_urls: list[HttpUrl] = Field(
-            default_factory=list, description="List of URLs"
-        )
-        list_of_urls_or_none: list[HttpUrl] | None = Field(
-            default=None, description="List of URLs or None"
-        )
-
-    # when
-    model = create_response_model(MyModel)
-    response = model(
-        score=0.9,
-        name="John Doe",
-        profile_url="http://example.com/johndoe",
-        list_of_urls=["http://example.com/url1", "http://example.com/url2"],
-    )
-
-    # then
-    assert "profile_url" not in response.model_dump()
-    assert "list_of_urls" not in response.model_dump()
-    assert "list_of_urls_or_none" not in response.model_dump()
